+++ designer_v2/lib/features/analyze/study_export_zip.dart
@@ -21,7 +21,11 @@ extension StudyExportZipX on StudyExportData {
     };
 
     files.forEach((filename, content) {
-      final archiveFile = ArchiveFile.string(filename, content);
+      final archiveFile = ArchiveFile.string(
+        filename,
+        // todo sanitize contents manually until archive v4 is released
+        content.replaceAll('’', "'").replaceAll('…', '...'),
+      );
       archive.addFile(archiveFile);
     });