+++ designer_v2/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.9.2
+
+ - **FIX**: upgrade deps.
+ - **FIX**: upgrade fvm to flutter v3.24.3.
+
+## 1.9.1
+
+ - **FIX**: upgrade to flutter v3.24.2.
+ - **FIX**: hotfix to sanitize export contents.
+
 ## 1.9.0
 
  - **FIX**: legend spacing.