+++ flutter_common/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.8.9
+
+ - **FIX**: upgrade deps.
+ - **FIX**: upgrade fvm to flutter v3.24.3.
+
+## 1.8.8
+
+ - **FIX**: upgrade to flutter v3.24.2.
+
 ## 1.8.7
 
  - **FIX**: upgrade deps.